+++ fuzz/fuzz_targets/fuzz_filter_summary.rs
@@ -1,0 +1,6 @@
+#![no_main]
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|data: &[u8]| {
+    let _ = zeppelin::index::ivf_flat::filter_summary::FilterCardinalitySummary::from_bytes(data);
+});